% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_monyr_means.R
\name{get_monyr_means}
\alias{get_monyr_means}
\title{Calculate mean air pollution levels by a specific time interval}
\usage{
get_monyr_means(df, time_unit = c("month", "year"))
}
\arguments{
\item{time_unit}{character; averaging interval ('prenatal', 'month", 'year')}

\item{d}{a data frame created by \code{prepdata()}}
}
\value{
a data frame
}
\description{
Calculate mean air pollution levels by a specific time interval
}
\examples{
\dontrun{

get_monyr_means(df, time_unit = 'month')
}

}
