% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prenatal_means.R
\name{get_prenatal_means}
\alias{get_prenatal_means}
\title{Get prenatal functions}
\usage{
get_prenatal_means(df)
}
\arguments{
\item{df}{a data frame created by \code{prepdata()}.}
}
\value{
a data frame with mean pollution measurements for prenatal intervals.
}
\description{
Get prenatal functions
}
\examples{
\dontrun{

get_prenatal_means(df)
}
}
