% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepdata.R
\name{prepdata}
\alias{prepdata}
\title{Clean and prepare daily air pollution data}
\usage{
prepdata(d, cohort_name = NULL)
}
\arguments{
\item{d}{a data frame with daily air pollution measurements}

\item{cohort}{character; name of cohort or study}
}
\value{
a data frame
}
\description{
Clean and prepare daily air pollution data
}
\examples{

\dontrun{
prepdata(d, cohort = NULL)
}

}
