% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_daily_pollution.R
\name{read_daily_pollution}
\alias{read_daily_pollution}
\title{Read daily air pollution}
\usage{
read_daily_pollution(path, n_lines = NULL)
}
\arguments{
\item{path}{character; folder path of input data frame}

\item{n_lines}{numeric; number of lines of input data frame to read.  This is useful to preview the data.}
}
\value{

}
\description{
\code{read_daily_pollution} can read both .CSV and .xlsx formats.  Date columns for
\code{dob} and \code{date} are automatically parsed.  If participant ID, DOB, and measurement
date columns are not automatically recognized, the user will be prompted to select
the corresponding variables in the input data frame.
}
\details{
columns are not present
}
\examples{
\dontrun{
read_daily_pollution(path, n_lines = NULL)
}

}
