% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_averages.R
\name{run_averages}
\alias{run_averages}
\title{Execute averaging script by specific interval}
\usage{
run_averages(
  input_path = NULL,
  export_path = NULL,
  average_type = "all",
  cohort = NULL,
  na_SAS = FALSE
)
}
\arguments{
\item{input_path}{character; path of input data frame}

\item{export_path}{character; path used to export output data}

\item{average_type}{character; one or combination of 'month', 'year', 'prenatal', or 'all'.
Default = 'all'.}

\item{cohort}{character; cohort name}

\item{na_SAS}{logical; if TRUE, missing values will be replaced with '.' instead of 'NA.'
Default = FALSE.}
}
\description{
Execute averaging script by specific interval
}
\examples{
\dontrun{

run_averages(input_path = NULL, export_path = NULL, average_type = 'all', cohort = NULL)
}
}
