% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mean_pollution.R
\name{get_mean_pollution}
\alias{get_mean_pollution}
\title{Calculate mean air pollution levels by a specific time interval}
\usage{
get_mean_pollution(d, time_unit = c("month", "year"))
}
\arguments{
\item{d}{a data frame created by \code{prepdata()}}

\item{time_unit}{character; averaging interval ('prenatal', 'month", 'year')}
}
\value{
a data frame
}
\description{
Calculate mean air pollution levels by a specific time interval
}
\examples{
\dontrun{

get_mean_pollution(d, time_unit = 'month')
}

}
